Add release trigger to GHA
Run the qcthis.yaml GHA when a release is released, and attach the report to that release as a `.txt` file. I also added code to pass `...` along to the `transform` function called in `format`, to allow me to print to a specific connection (ie a file).

I also cleaned up the pkgdown.yaml while I was at it (mostly matching gsm.core).

- Closes #71.