Add comments to PRs that trigger qc reports
Similar to how reports are attached to released as txt files, attach reports to PRs as comments. Make sure they aren't TOO annoying.

I also tagged all tests to specific issues. I'm not sure this will be the normal thing to do, but I prefer it to random floaters.

I also added a test for #47 (and refactored a little to enable it).

- Closes #72.