Add comments to PRs that trigger qc reports (#79)
* Add comments to PRs that trigger qc reports

Similar to how reports are attached to released as txt files, attach reports to PRs as comments. Make sure they aren't TOO annoying.

I also tagged all tests to specific issues. I'm not sure this will be the normal thing to do, but I prefer it to random floaters.

I also added a test for #47 (and refactored a little to enable it).

- Closes #72.

* Use first line of report as the summary for details block, and last line as a header.

* Add a header on the comment.

* Move footer to footer.

* Apply copilot suggestions.

I'll also manually trigger the workflow after this push to make sure the input works.