Make it easy to add qc report
Finalize qcthat-completed_issues.yaml script and add a function to enable it in a package. This involved refactoring a bunch of things and adding a convenience function to run everything (`QCPackage()`, #46) and a function to run the Completed Issues report specifically (`QCCompletedIssues()`, #80).

Note: The link in the qcthat-completed_issues.yaml files (both the one used in this package and the one available in `inst`) is a lie until this is released. I think I prefer to hard-code that and then update when we (plan to) release changes, like I did here.

- Closes #46
- Closes #73
- Closes #80