Make it easy to add qc report (#82)
* Make it easy to add qc report

Finalize qcthat-completed_issues.yaml script and add a function to enable it in a package. This involved refactoring a bunch of things and adding a convenience function to run everything (`QCPackage()`, #46) and a function to run the Completed Issues report specifically (`QCCompletedIssues()`, #80).

Note: The link in the qcthat-completed_issues.yaml files (both the one used in this package and the one available in `inst`) is a lie until this is released. I think I prefer to hard-code that and then update when we (plan to) release changes, like I did here.

- Closes #46
- Closes #73
- Closes #80

* Fix typos. Add "gh" to wordlist.

There's also a weird space character (via emoji) that we might have to deal with in the long run.

* Better report comment title.

And link to qcthat pkgdown.

* Put link in header of PR comment

* Added an argument to `print()` for IssueTestMatrix objects, `lglShowMilestones`, which allows the user to print without milestones.

Closes #40.

* Tag the requirement issue in tests.

* Repair action copier.