Optionally exclude qcthat-nocov labelled issues
Add an argument to several functions, `chrIgnoredLabels`, to determine whether issues with the `qcthat-nocov` label should be included in the report. Also add an argument to the `print()` and `format()` method for `qcthat_IssueTestMatrix`, `lglShowIgnoredLabels`, to control the display of a count of ignore issues by label.

I also updated the footer summaries to include counts, rather than "at least one".

- Closes #67
- Closes #90