Fix 67 90 Ignore Issues with `qcthat-nocov` label (#93)
* Resolve conflicts

* Rearrange files for cleaner & clearer navigation.

* Optionally exclude qcthat-nocov labelled issues

Add an argument to several functions, `chrIgnoredLabels`, to determine whether issues with the `qcthat-nocov` label should be included in the report. Also add an argument to the `print()` and `format()` method for `qcthat_IssueTestMatrix`, `lglShowIgnoredLabels`, to control the display of a count of ignore issues by label.

I also updated the footer summaries to include counts, rather than "at least one".

- Closes #67
- Closes #90

* Use `lglShowIgnoredLabels`

* Show proper verb in skip reports.

* Try to get it to show ignored issues again.

* Remove extra space

* TBD: Should I include the ignored footer by default, or hide by default?

* Redocument to fix ghosts.

* Clarify difference between `QCMergeGH()`, `QCMergeLocal()`, and `QCPR()`.

* Fix example reflow.

* Deal with test errors from version mismatch

Also protected against them somewhat in the future, so the report works (and shows the error) rather than failing entirely.